+++ CHANGELOG.md
@@ -12,8 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added the functionality to show/hide and lock all regions ([#1796](https://github.com/cartavis/carta-frontend/issues/1796)).
 * Added a method to auto-scrolling the selected region into the region list view ([#1797](https://github.com/CARTAvis/carta-frontend/issues/1797)).
 * Added the functionality to mirror cursor position on spatially matched frame via hotkey "G" ([#1947](https://github.com/cartavis/carta-frontend/issues/1947)).
-* Added buttons to image fitting for exporting fitting result and full log ([#1397](https://github.com/cartavis/carta-frontend/issues/1397)).
-* Added support for image fitting with fixed parameters ([#1397](https://github.com/CARTAvis/carta-frontend/issues/1397)).
+* Added support for fitting images with fixed parameters, generating model and residual images, and exporting fitting result and full log; added support for estimating progress and cancelling tasks ([#1397](https://github.com/CARTAvis/carta-frontend/issues/1397)).
 * Added tooltip to the Image column for Image List and Cursor Info overlay ([#1948](https://github.com/CARTAvis/carta-frontend/issues/1948)).
 * Added additional cursor info option to spectral profile widget ([#1837](https://github.com/CARTAvis/carta-frontend/issues/1837)).
 * Added a selection option in the PV generator widget to swap x and y axis, an input for spectral axis limit, and a toggle button to let users decide whether or not to keep the previously generated PV images ([#1950](https://github.com/cartavis/carta-frontend/issues/1950), [#1951](https://github.com/cartavis/carta-frontend/issues/1951), [#1952](https://github.com/cartavis/carta-frontend/issues/1952)).